third_party/blink/web_tests/platform/generic/fast/webgl/renderer-and-vendor-strings-expected.txt

Verifies the contents of the RENDERER and VENDOR strings to avoid regressions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl.getParameter(gl.RENDERER) is "WebKit WebGL"
PASS gl.getParameter(gl.VENDOR) is "WebKit"
PASS successfullyParsed is true

TEST COMPLETE
